% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/incidence_by_quint.R
\name{incidence_by_quint}
\alias{incidence_by_quint}
\title{Calculate program incidence by income quintile}
\usage{
incidence_by_quint(df, program_var, inc_var, weight_var)
}
\arguments{
\item{df}{Household-level dataset containing a variable for program receipt, income and survey weights, dataframe}

\item{program_var}{Name of binary variable which indicates whether household received program, character}

\item{inc_var}{Name of income variable, character}

\item{weight_var}{Name of survey weight variable, character}
}
\description{
Calculate program incidence by income quintile
}
